// models/DisplayOptionsEnum.ts
export enum Featured {
  featured = 'featured',
  notFeatured = 'not featured',
  withInDateRange = 'feature item with in date range'
}

export enum DisplayItemOptions {
  hide = 'hide',
  display = 'display',
  withInDateRange = 'display item within date range'
}

export enum DisplayOptionsPostType {
  template = 'template',
  page = 'page',
  staticDocument = 'static document',
  externalUrl = 'external url'
}